+++ src/finlang/ffiAttribute.cs
@@ -1,0 +1,14 @@
+﻿#pragma warning disable IDE1006 // Naming Styles
+
+using System;
+
+namespace finlang;
+
+/// <summary>
+/// Foreign Function Interface. This attribute is used to mark a method as a foreign function.
+/// Classes marked with this attribute will NOT be transpiled to C99.
+/// </summary>
+[AttributeUsage(AttributeTargets.Class)]
+public class ffiAttribute : Attribute {
+
+}